refactor: seperate characper provider from simulation setting
