+++ tests/simulate/policy/test_dsl_runtime.py
@@ -3,11 +3,8 @@ from pathlib import Path
 import pytest
 
 import simaple.simulate.component.skill  # noqa: F401
-from simaple.container.simulation import (
-    SimulationContainer,
-    SimulationSetting,
-    BaselineSimulationConfig,
-)
+from simaple.container.character_provider import BaselineSimulationConfig
+from simaple.container.simulation import SimulationContainer, SimulationSetting
 from simaple.core.job_category import JobCategory
 from simaple.core.jobtype import JobType
 from simaple.simulate.policy.parser import parse_dsl_to_operations
